fix showing the right navbar to users, also hiding company pages from participants and vice versa
